% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations_classify.R
\name{classify_occupation}
\alias{classify_occupation}
\title{Classify occupations}
\usage{
classify_occupation(
  corpus,
  id_col = "id",
  text_col = "text",
  lang = "en",
  num_leaves = 10,
  isco_level = 3,
  string_dist = NULL
)
}
\arguments{
\item{corpus}{A data.frame or a data.table that contains the id and the text variables.}

\item{id_col}{The name of the id variable.}

\item{text_col}{The name of the text variable.}

\item{lang}{The language that the text is in.}

\item{num_leaves}{The number of occupations/neighbors that are kept when matching.}

\item{isco_level}{The \href{https://ec.europa.eu/esco/portal/escopedia/Occupation}{ISCO} level of the suggested occupations.
Can be either 1, 2, 3, 4 for ISCO occupations, or NULL for ESCO occupations.}

\item{string_dist}{String distance used for fuzzy matching. The \link[stringdist]{amatch} function from the stringdist package is used.}
}
\value{
Either a data.table with the id, the preferred label and the suggested ESCO occupation URIs (num_leaves predictions for each id),
or a data.table with the id, the preferred label and the suggested ISCO group of the inputted level (one for each id).
}
\description{
This function takes advantage of the hierarchical structure of the ESCO-ISCO mapping and matches multilingual free-text with the
\href{https://ec.europa.eu/esco/portal/home}{ESCO} occupations vocabulary in order to map semi-structured vacancy data into the official
ESCO-ISCO classification.
}
\details{
First, the input text is cleansed and tokenized. The tokens are then matched with the ESCO occupations vocabulary, created from
the \link[=occupations_bundle]{preferred and alternative labels} of the occupations, and joined with the \link[=tf_idf]{tfidf}
weighted tokens of the ESCO occupations. After the terms are joined, the suggested ESCO occupations are retrieved using a weighted sum model.
If a ISCO level was specified, the K-Nearest Neighbors algorithm is used to determine the suggested occupation of the requested ISCO level,
classified by a plurality vote of its neighbors.

Before the suggestions are returned, the preferred label of each suggested occupation is added to the result, using the
\code{\link{occupations_bundle}} and \code{\link{isco_occupations_bundle}} as look-up tables.
}
\examples{
corpus <- data.frame(
 id = 1:3,
 text = c("Data Scientist", "Junior Architect Engineer", "Cashier at McDonald's")
)
classify_occupation(corpus = corpus, isco_level = NULL, lang = "en", num_leaves = 5)

}
\references{
M.P.J. van der Loo (2014). \href{https://journal.r-project.org/archive/2014-1/loo.pdf}{The stringdist package for approximate string matching}. R Journal 6(1) pp 111-122.

Turrell, A., Speigner, B., Djumalieva, J., Copple, D., and Thurgood, J. (2018).
\href{https://www.bankofengland.co.uk/working-paper/2018/using-job-vacancies-to-understand-the-effects-of-labour-market-mismatch-on-uk-output}{
Using job vacancies to understand the effects of labour market mismatch on UK output and productivity}, Staff Working Paper 737, Bank of England.

ESCO Service Platform - The ESCO Data Model \href{https://ec.europa.eu/esco/portal/document/en/87a9f66a-1830-4c93-94f0-5daa5e00507e}{documentation}
}
