% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_stopwords}
\alias{get_stopwords}
\title{Retrieve stopwords}
\usage{
get_stopwords(code)
}
\arguments{
\item{code}{A string with the language code of the stopwords.}
}
\value{
Character vector with the stopwords or NULL if the language code is unknown.
}
\description{
The functions retrieves stopwords from the \code{\link{stopwords}} package using the ISO-639-1 encoding.
For miscellaneous languages \code{\link{data_stopwords_misc}} are used.
}
\examples{
get_stopwords("en")[1:10]

}
