% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My Hello}
\usage{
hello(x)
}
\arguments{
\item{x}{The name}
}
\value{
The output \code{\link{print}} and of \code{\link{print}}
}
\description{
My Hello
}
\examples{
hello("aaaa")
\dontrun{
hello("CoRona")
}
}
