% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_iris.R
\docType{data}
\name{tib_iris}
\alias{tib_iris}
\title{Sample iris tibble}
\format{A tibble with 5 elements, which are:
\describe{
\item{Sepal.Length}{The sepal length of the species}
\item{Sepal.Width}{The sepal length of the species}
\item{Petal.Length}{The sepal length of the species}
\item{Petal.Width}{The sepal length of the species}
\item{Species}{The sepal length of the \code{Species}}
}}
\usage{
tib_iris
}
\description{
A data set containing the iris data stored in a tibble
}
\keyword{datasets}
