% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupations_classify.R
\name{classify_occupation}
\alias{classify_occupation}
\title{Classify occupations}
\usage{
classify_occupation(
  corpus,
  id_col = "id",
  text_col = "text",
  lang = "en",
  num_leaves = 10,
  isco_level = 3,
  max_dist = 0.1,
  string_dist = NULL
)
}
\arguments{
\item{corpus}{A data.frame or a data.table that contains the id and the text variables.}

\item{id_col}{The name of the id variable.}

\item{text_col}{The name of the text variable.}

\item{lang}{The language that the text is in.}

\item{num_leaves}{The number of occupations/neighbors that are kept when matching.}

\item{isco_level}{The \href{https://ec.europa.eu/esco/portal/escopedia/Occupation}{ISCO} level of the suggested occupations.
Can be either 1, 2, 3, 4 for ISCO occupations, or NULL that returns ESCO occupations.}

\item{max_dist}{String distance used for fuzzy matching. The \code{\link[stringdist]{amatch}} function from the stringdist package is used.}

\item{string_dist}{String dissimilarity measurement. Available string distance metrics: \code{\link[stringdist]{stringdist-metrics}}.}
}
\value{
Either a data.table with the id, the preferred label and the suggested ESCO occupation URIs (num_leaves predictions for each id),
or a data.table with the id, the preferred label and the suggested ISCO group of the inputted level (one for each id).
}
\description{
This function takes advantage of the hierarchical structure of the ESCO-ISCO mapping and matches multilingual free-text with the
\href{https://ec.europa.eu/esco/portal/home}{ESCO} occupations vocabulary in order to map semi-structured vacancy data into the official
ESCO-ISCO classification.
}
\details{
First, the input text is cleansed and tokenized. The tokens are then matched with the ESCO occupations vocabulary, created from
the \code{\link[=occupations_bundle]{preferred and alternative labels}} of the occupations. They are joined with the \code{\link[=tf_idf]{tfidf}}
weighted tokens of the ESCO occupations and the sum of the tf-idf score is used to retrieve the suggested ontologies. Technically speaking, the
suggested ESCO occupations are retrieved by solving the optimization problem, \deqn{\arg\max_d\left\{\vec{u}_{binary}\cdot \vec{u}_d\right\}}
where, \eqn{\vec{u}_{binary}} stands for the binary representation of a query to the ESCO-vocabulary space,
while, \eqn{\vec{u}_d} is the ESCO occupation normalized vector generated by the tf-idf numerical statistic.
If an ISCO level is specified, the K-NN algorithm is used to determine the suggested occupation, classified by a plurality vote in the
corresponding hierarchical level of its neighbors.

Before the suggestions are returned, the preferred label of each suggested occupation is added to the result, using the
\code{\link{occupations_bundle}} and \code{\link{isco_occupations_bundle}} as look-up tables.
}
\examples{
corpus <- data.frame(
 id = 1:3,
 text = c(
   "Junior Architect Engineer",
   "Cashier at McDonald's",
   "Priest at St. Martin Catholic Church"
 )
)
classify_occupation(corpus = corpus, isco_level = 3, lang = "en", num_leaves = 5)

}
\references{
M.P.J. van der Loo (2014). \href{https://journal.r-project.org/archive/2014-1/loo.pdf}{The stringdist package for approximate string matching}. R Journal 6(1) pp 111-122.

Turrell et al., (2018).
\href{https://www.bankofengland.co.uk/working-paper/2018/using-job-vacancies-to-understand-the-effects-of-labour-market-mismatch-on-uk-output}{
Using job vacancies to understand the effects of labour market mismatch on UK output and productivity}, Staff Working Paper 737, Bank of England.

ESCO Service Platform - The ESCO Data Model \href{https://ec.europa.eu/esco/portal/document/en/87a9f66a-1830-4c93-94f0-5daa5e00507e}{documentation}
}
