% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansing_corpus.R
\name{cleansing_corpus}
\alias{cleansing_corpus}
\title{Title}
\usage{
cleansing_corpus(
  text,
  rem_newline = TRUE,
  rem_nonalphanum = TRUE,
  rem_longwords = TRUE,
  rem_space = TRUE,
  tolower = TRUE
)
}
\arguments{
\item{text}{character vector}

\item{rem_newline}{boolean}

\item{rem_nonalphanum}{boolean}

\item{rem_longwords}{boolean}

\item{rem_space}{boolean}

\item{tolower}{boolean}
}
\value{
A character vector
}
\description{
Title
}
\examples{
txt <- "It has roots in a piece of classical Latin literature from 45 BC"
cleansing_corpus(txt)
}
