% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_language.R
\name{identify_language}
\alias{identify_language}
\title{Detect Language}
\usage{
identify_language(text)
}
\arguments{
\item{text}{A string with text to classify or a connection to read from.

\itemize{
\item cld2: Probabilistically (Naïve Bayesian classifier) detects over 80 languages in plain text.
}}
}
\value{
A character vector with ISO-639-1 two-letter language codes.
}
\description{
This function performs language detection by using Compact Language Detector 2 from CRAN library \code{\link{cld2}}.
It is vectorised and guesses the language of each string. Note that it is not designed to do well on very short text,
lists of proper names, part numbers, etc. CLD2 has the highest F1 score and is an order of magnitude faster than CLD3.
}
\examples{
txt <- c("English is a West Germanic language ", "In espaniol, le lingua castilian")
identify_language(txt)

}
